% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_product.R
\name{check_product}
\alias{check_product}
\title{Local arcgisbinding::arc.check_product() function.}
\usage{
check_product()
}
\value{
The output of arcgisbinding::arc.check_product()
}
\description{
This is a copy of the arcgisbinding::arc.check_product() function created
to avoid errors in R CMD Check's on GitHub. Since arcgisbinding is a windows
binary package, the R CMD Check resolver fails in loading the package.
}
