% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_arcpy.R
\name{init_arcpy}
\alias{init_arcpy}
\title{Initialize Connection to ArcGIS and \code{arcpy}}
\usage{
init_arcpy(python_env_path = NULL, conda_env = NULL)
}
\arguments{
\item{python_env_path}{Requires a string to the python executable path for the \code{arcpy} environment.}

\item{conda_env}{Requires a string of the conda environment name containing the \code{arcpy} environment.}
}
\value{
A global variable called \code{arcpy} loaded as a python module and the output of
arcgisbinding::arc.check_product() detailing ESRI product information.
}
\description{
The function initializes a connection to either ArcGIS Pro or ArcMap and also loads
the \code{arcpy} python module into the global environment.
}
\examples{
\dontrun{
init_arcpy(python_env_path = "C:/ESRI/conda/envs/{conda environment name}/python.exe")
init_arcpy(conda_env_path = "{conda environment name}")
# product: ArcGIS Pro (13.1.0.41833)
# license: Advanced
# version: 1.0.1.300
}
}
