% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_arcpy.R
\name{init_arcpy}
\alias{init_arcpy}
\title{Initialize Connection to ArcGIS}
\usage{
init_arcpy(python_env_path = NULL, conda_env = NULL)
}
\arguments{
\item{python_env_path}{This is a string to the python executable path in the \code{arcpy} environment.}

\item{conda_env}{This is a string of the conda environment name for the \code{arcpy} environment.}
}
\value{
The path or environment selected for arcpy and a description of your ArcGIS product using arcgisbinding::arc.check_product().
}
\description{
Initialize Connection to ArcGIS
}
\examples{
\dontrun{
init_arcpy(python_env_path = "C:/ESRI/conda/envs/{conda environment name}/python.exe")
init_arcpy(conda_env_path = "{conda environment name}")
# product: ArcGIS Pro (13.1.0.41833)
# license: Advanced
# version: 1.0.1.300
}
}
